+++ CHANGELOG.md
@@ -50,6 +50,7 @@
 * [BUGFIX] Add `__markers__` tenant ID validation. #6761
 * [BUGFIX] Ring: Fix nil pointer exception when token is shared. #6768
 * [BUGFIX] Fix race condition in active user. #6773
+* [BUGFIX] Ingester: Allow shipper to skip corrupted blocks. #6786
 
 ## 1.19.0 2025-02-27